chore(solver/app): enable swaps on mainnet
